Add stage 8/8 of "Smart Calculator"

+++ Smart Calculator/Smart Calculator/task/src/calculator/Main.kt
@@ -1,6 +1,6 @@
 package calculator
 
-import kotlin.math.pow
+import java.math.BigInteger
 import com.github.h0tk3y.betterParse.combinators.*
 import com.github.h0tk3y.betterParse.grammar.Grammar
 import com.github.h0tk3y.betterParse.grammar.parser
@@ -10,7 +10,7 @@ import com.github.h0tk3y.betterParse.lexer.regexToken
 import com.github.h0tk3y.betterParse.parser.Parser
 
 sealed interface CalcExpression
-data class Num(val n: Int): CalcExpression
+data class Num(val n: BigInteger): CalcExpression
 data class Nme(val s: String): CalcExpression
 data class Neg(val e: CalcExpression): CalcExpression
 data class Pow(val e1: CalcExpression, val e2: CalcExpression): CalcExpression
@@ -22,11 +22,11 @@ data class Bnd(val s: String, val e: CalcExpression): CalcExpression
 
 class EvalException(message: String): Exception(message)
 
-fun eval(e: CalcExpression, bnds: Map<String, Int>): Pair<Int?, Map<String, Int>> = when (e) {
+fun eval(e: CalcExpression, bnds: Map<String, BigInteger>): Pair<BigInteger?, Map<String, BigInteger>> = when (e) {
     is Num -> e.n to bnds
     is Nme -> if (bnds.containsKey(e.s)) bnds[e.s] to bnds else throw EvalException("Unknown variable")
     is Neg -> -eval(e.e, bnds).first!! to bnds
-    is Pow -> eval(e.e1, bnds).first!!.toDouble().pow(eval(e.e2, bnds).first!!.toDouble()).toInt() to bnds
+    is Pow -> eval(e.e1, bnds).first!!.pow(eval(e.e2, bnds).first!!.toInt()) to bnds
     is Mul -> eval(e.e1, bnds).first!! * eval(e.e2, bnds).first!! to bnds
     is Div -> eval(e.e1, bnds).first!! / eval(e.e2, bnds).first!! to bnds
     is Add -> eval(e.e1, bnds).first!! + eval(e.e2, bnds).first!! to bnds
@@ -52,7 +52,7 @@ class CalcGrammar: Grammar<CalcExpression>() {
     private val wsp by regexToken("\\s+", ignore = true)
 
     private val term: Parser<CalcExpression> by
-        (num use { Num(text.toInt()) }) or
+        (num use { Num(BigInteger(text)) }) or
         (nme use { Nme(text) }) or
         (skip(min) and parser(::term) map { Neg(it) }) or
         (skip(lpr) and parser(::rootParser) and skip(rpr))
@@ -94,7 +94,7 @@ fun preParse(input: String): String {
     return input
 }
 
-fun repl(grmr: CalcGrammar = CalcGrammar(), bnds: Map<String, Int> = emptyMap()) {
+fun repl(grmr: CalcGrammar = CalcGrammar(), bnds: Map<String, BigInteger> = emptyMap()) {
     val input = readln()
     if ((input.startsWith("/"))) {
         when (input.drop(1)) {
